@@ -2,4 +2,24 @@
 
 __author__ = """Andrew Chen Wang"""
 __email__ = "acwangpython@gmail.com"
-__version__ = "0.1.0"
+
+VERSION = (0, 2, 0)
+__version__ = ".".join(map(str, VERSION))
+
+
+def get_redis_connection(alias="default", write=True):
+    """
+    Helper used for obtaining a raw redis client.
+    """
+
+    from django.core.cache import caches
+
+    cache = caches[alias]
+
+    if not hasattr(cache, "client"):
+        raise NotImplementedError("This backend does not support this feature")
+
+    if not hasattr(cache.client, "get_client"):
+        raise NotImplementedError("This backend does not support this feature")
+
+    return cache.client.get_client(write)
