@@ -2,8 +2,8 @@ class BaseCompressor:
     def __init__(self, options):
         self._options = options
 
-    def compress(self, value):
+    def compress(self, value: bytes) -> bytes:
         raise NotImplementedError
 
-    def decompress(self, value):
+    def decompress(self, value: bytes) -> bytes:
         raise NotImplementedError
