@@ -2,8 +2,8 @@ from .base import BaseCompressor
 
 
 class IdentityCompressor(BaseCompressor):
-    def compress(self, value):
+    def compress(self, value: bytes) -> bytes:
         return value
 
-    def decompress(self, value):
+    def decompress(self, value: bytes) -> bytes:
         return value
