@@ -1,10 +1,16 @@
-from typing import Any, Dict
+from typing import TYPE_CHECKING, Any, Dict
+from urllib.parse import parse_qs, urlparse
 
-from aioredis import Redis
-from aioredis.parser import Reader
-from aioredis.pool import ConnectionsPool
 from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 from django.utils.module_loading import import_string
+from redis.asyncio import Redis
+from redis.asyncio.connection import DefaultParser, to_bool
+from redis.asyncio.sentinel import Sentinel
+
+
+if TYPE_CHECKING:
+    from redis.asyncio import ConnectionPool
 
 
 class ConnectionFactory:
@@ -15,115 +21,161 @@ class ConnectionFactory:
     # ConnectionFactory is instantiated, as Django creates new cache client
     # (DefaultClient) instance for every request.
 
-    _pools = {}
+    _pools: Dict[str, "ConnectionPool"] = {}
 
     def __init__(self, options):
-        # Pool
         pool_cls_path = options.get(
-            "CONNECTION_POOL_CLS", "aioredis.pool.ConnectionsPool"
+            "CONNECTION_POOL_CLASS", "redis.asyncio.connection.ConnectionPool"
         )
-        # Must subclass AbcPool
-        self.pool_cls: ConnectionsPool = import_string(pool_cls_path)
+        self.pool_cls = import_string(pool_cls_path)
+        self.pool_cls_kwargs = options.get("CONNECTION_POOL_KWARGS", {})
 
-        pool_func_path = options.get("CONNECTION_POOL_FUNC", "aioredis.create_pool")
-        self.pool_func = import_string(pool_func_path)
-        self.pool_func_kwargs: Dict[str, Any] = options.get(
-            "CONNECTION_POOL_KWARGS", {}
+        redis_client_cls_path = options.get(
+            "REDIS_CLIENT_CLASS", "redis.asyncio.client.Redis"
         )
-
-        # Client
-        redis_client_func_path = options.get(
-            "REDIS_CLIENT_FUNC", "aioredis.create_redis"
-        )
-        self.redis_client_func = import_string(redis_client_func_path)
-        self.redis_client_func_kwargs = options.get("REDIS_CLIENT_KWARGS", {})
-        # Must subclass AbcConnection
-        redis_client_cls_path = options.get("REDIS_CLIENT_CLASS", "aioredis.Redis")
         self.redis_client_cls = import_string(redis_client_cls_path)
+        self.redis_client_cls_kwargs = options.get("REDIS_CLIENT_KWARGS", {})
 
         self.options = options
 
-    def make_connection_params(self, address) -> Dict[str, Any]:
+    def make_connection_params(self, url: str) -> Dict[str, Any]:
         """
         Given a main connection parameters, build a complete
         dict of connection parameters.
         """
 
         kwargs = {
-            "address": address,
-            "parser": self.get_parser_cls(),
+            "url": url,
+            "parser_class": self.get_parser_cls(),
         }
 
         password = self.options.get("PASSWORD", None)
         if password:
             kwargs["password"] = password
 
-        create_connection_timeout = self.options.get("CREATE_CONNECTION_TIMEOUT", None)
-        if create_connection_timeout:
+        socket_timeout = self.options.get("SOCKET_TIMEOUT", None)
+        if socket_timeout:
             assert isinstance(
-                create_connection_timeout, (int, float)
-            ), "Create connection timeout should be float or integer"
-            assert (
-                create_connection_timeout > 0
-            ), "Create connection timeout should be greater than 0"
-            kwargs["create_connection_timeout"] = create_connection_timeout
+                socket_timeout, (int, float)
+            ), "Socket timeout should be float or integer"
+            kwargs["socket_timeout"] = socket_timeout
+
+        socket_connect_timeout = self.options.get("SOCKET_CONNECT_TIMEOUT", None)
+        if socket_connect_timeout:
+            assert isinstance(
+                socket_connect_timeout, (int, float)
+            ), "Socket connect timeout should be float or integer"
+            kwargs["socket_connect_timeout"] = socket_connect_timeout
 
         return kwargs
 
-    async def connect(self, address: str) -> Redis:
+    async def connect(self, url: str) -> Redis:
         """
         Given a basic connection parameters,
         return a new connection.
         """
-        params: Dict[str, Any] = self.make_connection_params(address)
+        params: Dict[str, Any] = self.make_connection_params(url)
         return await self.get_connection(params)
+
+    async def disconnect(self, connection: Redis):
+        """
+        Given a not null client connection it disconnect from the Redis server.
+        The default implementation uses a pool to hold connections.
+        """
+        await connection.connection_pool.disconnect()
 
     async def get_connection(self, params: Dict[str, Any]) -> Redis:
         """
-        Given pre-formatted params, return a
+        Given a now preformatted params, return a
         new connection.
-
         The default implementation uses a cached pools
         for create new connection.
         """
-        # Creates new connection if needed. If pool is closed, womp womp
-        pool: ConnectionsPool = await self.get_or_create_connection_pool(params)
-        # If no free connection, waits for another conn.
-        # Raises RunTime error if we run out of connections by maxsize
-        return self.redis_client_cls(pool)
+        pool = self.get_or_create_connection_pool(params)
+        return await self.redis_client_cls(
+            connection_pool=pool, **self.redis_client_cls_kwargs
+        )
 
     def get_parser_cls(self):
         cls = self.options.get("PARSER_CLASS", None)
         if cls is None:
-            return Reader
+            return DefaultParser
         return import_string(cls)
 
-    async def get_or_create_connection_pool(
-        self, params: Dict[str, Any]
-    ) -> ConnectionsPool:
+    def get_or_create_connection_pool(self, params: Dict[str, Any]):
         """
-        Given connection parameters and return a new
+        Given a connection parameters and return a new
         or cached connection pool for them.
-
         Reimplement this method if you want distinct
         connection pool instance caching behavior.
         """
-        key = params["address"]
+        key = params["url"]
         if key not in self._pools:
-            self._pools[key] = await self.get_connection_pool(params)
+            self._pools[key] = self.get_connection_pool(params)
         return self._pools[key]
 
-    async def get_connection_pool(self, params: Dict[str, Any]) -> ConnectionsPool:
+    def get_connection_pool(self, params: Dict[str, Any]):
         """
-        Given connection parameters, return a new
+        Given a connection parameters, return a new
         connection pool for them.
-
         Overwrite this method if you want a custom
         behavior on creating connection pool.
         """
         cp_params = dict(params)
-        cp_params.update(self.pool_func_kwargs)
-        pool: ConnectionsPool = await self.pool_func(**cp_params)
+        cp_params.update(self.pool_cls_kwargs)
+        pool: "ConnectionPool" = self.pool_cls.from_url(**cp_params)
+
+        if pool.connection_kwargs.get("password", None) is None:
+            pool.connection_kwargs["password"] = params.get("password", None)
+            pool.reset()
+
+        return pool
+
+
+class SentinelConnectionFactory(ConnectionFactory):
+    def __init__(self, options):
+        # allow overriding the default SentinelConnectionPool class
+        options.setdefault(
+            "CONNECTION_POOL_CLASS", "redis.asyncio.sentinel.SentinelConnectionPool"
+        )
+        super().__init__(options)
+
+        sentinels = options.get("SENTINELS")
+        if not sentinels:
+            raise ImproperlyConfigured(
+                "SENTINELS must be provided as a list of (host, port)."
+            )
+
+        # provide the connection pool kwargs to the sentinel in case it
+        # needs to use the socket options for the sentinels themselves
+        connection_kwargs = self.make_connection_params(None)
+        connection_kwargs.pop("url")
+        connection_kwargs.update(self.pool_cls_kwargs)
+        self._sentinel = Sentinel(
+            sentinels,
+            sentinel_kwargs=options.get("SENTINEL_KWARGS"),
+            **connection_kwargs,
+        )
+
+    def get_connection_pool(self, params: Dict[str, Any]):
+        """
+        Given a connection parameters, return a new sentinel connection pool
+        for them.
+        """
+        url = urlparse(params["url"])
+
+        # explicitly set service_name and sentinel_manager for the
+        # SentinelConnectionPool constructor since will be called by from_url
+        cp_params = dict(params)
+        cp_params.update(service_name=url.hostname, sentinel_manager=self._sentinel)
+        pool = super().get_connection_pool(cp_params)
+
+        # convert "is_master" to a boolean if set on the URL, otherwise if not
+        # provided it defaults to True.
+        is_master = parse_qs(url.query).get("is_master")
+        if is_master:
+            pool.is_master = to_bool(is_master[0])
+
         return pool
 
 
