@@ -1,16 +1,25 @@
+import asyncio
 from sys import version_info
 
 import pytest
-from django.core.cache import cache
+import pytest_asyncio
+from django.core.cache import cache as default_cache
 
 from django_async_redis.cache import RedisCache
 
 
-@pytest.fixture(autouse=True)
-async def cache_setup() -> RedisCache:
-    await cache.clear_async()
-    yield cache
-    await cache.client.close_connections()
+@pytest.fixture(scope="session")
+def event_loop():
+    policy = asyncio.get_event_loop_policy()
+    loop = policy.new_event_loop()
+    yield loop
+    loop.close()
+
+
+@pytest_asyncio.fixture()
+async def cache() -> RedisCache:
+    yield default_cache
+    await default_cache.aclear()
 
 
 if version_info >= (3, 8):
